backend/requirements.txt
fastapi==0.95.1
uvicorn==0.23.2
pydantic==1.10.8
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.107.0
google-generativeai==0.3.2
python-multipart==0.0.6
httpx==0.25.1
python-dotenv==1.0.0
pytest==7.4.3
requests==2.31.0
jinja2==3.1.2
email-validator==2.1.0.post1